server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Soorya U <sooryau7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
